Fix references/paths in previous controllers/views.
