+++ src/main/java/com/codeup/adlister/controllers/HelloWorldServlet.java
@@ -7,12 +7,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "controllers.HelloWorldServlet", urlPatterns = "/")
+@WebServlet(name = "controllers.HelloWorldServlet", urlPatterns = "/hello")
 public class HelloWorldServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
         String name = request.getParameter("name");
 
-        if(name.isEmpty()){
+        if(name != null && name.isEmpty()){
             name = "world";
         }
