+++ src/main/java/com/codeup/adlister/controllers/UsersServlet.java
@@ -1,0 +1,28 @@
+package com.codeup.adlister.controllers;
+
+import com.codeup.adlister.dao.DaoFactory;
+import com.codeup.adlister.models.User;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.List;
+
+/**
+ * Created by Fer on 6/7/17.
+ */
+@WebServlet(name = "UsersServlet", urlPatterns = "/users")
+public class UsersServlet extends HttpServlet {
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+    }
+
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        List<User> users = DaoFactory.getUsersDao().all();
+        request.setAttribute("instructors", users);
+        request.getRequestDispatcher("/users.jsp").forward(request, response);
+    }
+}
