+++ src/main/webapp/hello.jsp
@@ -2,15 +2,14 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <html>
 <head>
-    <jsp:include page="partials/head.jsp">
+    <jsp:include page="/WEB-INF/partials/head.jsp">
         <jsp:param name="title" value="Welcome to my site!" />
     </jsp:include>
 </head>
 <body>
-<jsp:include page="partials/navbar.jsp" />
+<jsp:include page="/WEB-INF/partials/navbar.jsp" />
 <div class="container">
     <h1>Hello, ${name} ${lastName}</h1>
-
 </div>
 </body>
-</ht</>
+</html>
